requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
av==16.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.6.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.121.2
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
jiter==0.12.0
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.23.2
openai==2.8.0
packaging==25.0
pandas==2.3.3
protobuf==6.33.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
